refactor(adminPanel): NASS-1908: extract ‘Delete role’ modal
